Добавила новый модуль с функциями возвращения в исходное состояние
