+++ js/functions-to-reset.js
@@ -1,0 +1,37 @@
+import {CENTER_TOKIO_LAT, CENTER_TOKIO_LNG} from './const.js';
+import {MAP_ZOOM, map, mainPinMarker} from './map.js';
+
+const adForm = document.querySelector('.ad-form');
+const addressField = adForm.querySelector('[name="address"]');
+const resetButton = adForm.querySelector('.ad-form__reset');
+const mapFiltersForm = document.querySelector('.map__filters');
+
+// Возвращение карты в исходное состояние
+
+const resetMap = () => {
+  mainPinMarker.setLatLng({
+    lat: CENTER_TOKIO_LAT,
+    lng: CENTER_TOKIO_LNG,
+  });
+  map.setView({
+    lat: CENTER_TOKIO_LAT,
+    lng: CENTER_TOKIO_LNG,
+  }, MAP_ZOOM);
+
+  addressField.value = `${CENTER_TOKIO_LAT}, ${CENTER_TOKIO_LNG}`;
+};
+
+// Очистка формы
+
+const resetForm = () => {
+  adForm.reset();
+  mapFiltersForm.reset();
+  resetMap();
+};
+
+resetButton.addEventListener('click', (evt) => {
+  evt.preventDefault();
+  resetForm();
+});
+
+export {resetForm};
